Fix: remove BigInt, ensure emails fire after payment

+++ api/create-payment.js
@@ -142,7 +142,7 @@ module.exports = async function handler(req, res) {
       sourceId,
       idempotencyKey: orderId,
       amountMoney: {
-        amount: BigInt(Math.round(amount * 100)),
+        amount: Math.round(amount * 100),
         currency,
       },
       note: `Side Quest Complete — Order ${orderId}`,
